Functions

Reference shapes

Elfel.RefShapes.manifdim — Function

manifdim(rs)

Get the manifold dimension of the reference shape.

Elfel.RefShapes.manifdimv — Function

manifdimv(::Type{T}) where {T<:AbstractRefShape{MANIFDIM}} where {MANIFDIM}

Get the manifold dimension of the reference shape as Val.

Missing docstring.

Missing docstring for quadrature. Check Documenter's build log for details.

Elements

Elfel.FElements.shapedesc — Function

shapedesc(fe::FE{RS, SD}) where {RS, SD}

Topological shape description.

Refer to the MeshCore library.

Elfel.FElements.refshape — Function

refshape(fe::FE{RS, SD}) where {RS, SD}

Reference shape.

Elfel.FElements.nfeatofdim — Function

nfeatofdim(fe::FE{RS, SD}, m) where {RS, SD}

Number of features of manifold dimension m. Note that 0 <= m <= 3.

Elfel.FElements.ndofperfeat — Function

feathasdof(fe::FE{RS, SD}, m) where {RS, SD}

How many degrees of freedom are attached to a the feature of manifold dimension m?

Note that 0 <= m <= 3.

Elfel.FElements.ndofsperel — Function

ndofsperel(it::FEIterator)

Retrieve the number of degrees of freedom per element.

ndofsperel(fe::FE{RS, SD}) where {RS, SD}

Provide the number of degrees of freedom per element.

Enumerate all features of all manifold dimensions, and for each feature multiply by the number of degrees of freedom per feature. The assumption is that this is a scalar finite element.

MeshCore.manifdim — Function

manifdim(fe::FE{RS, SD}) where {RS, SD}

Get the manifold dimension of the finite element.

Elfel.FElements.Jacobian — Function

Jacobian(::Val{0}, J::T) where {T}

Evaluate the point Jacobian.

J = Jacobian matrix, which isn't really defined well for a 0-manifold.

Jacobian(::Val{1}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Jacobian(::Val{2}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Jacobian(self::T, J::FFltMat)::FFlt where {T<:AbstractFESet3Manifold}

Evaluate the volume Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Elfel.FElements.jacjac — Function

jacjac(it::QPIterator)

Compute the Jacobian matrix and the Jacobian determinant.

At the current integration point.

jacjac(fe::FE{RS, SD}, locs, gradNpar) where {RS, SD}

Compute the Jacobian matrix and the Jacobian determinant.

This is the generic version suitable for isoparametric elements.

Missing docstring.

Missing docstring for FEH1_L2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_Q4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T6. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T3_BUBBLE. Check Documenter's build log for details.

Fields

Missing docstring.

Missing docstring for doftype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dofnumtype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nterms. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ndofsperterm. Check Documenter's build log for details.

Elfel.FEFields.ndofs — Function

ndofs(fesp::FES)  where {FES<:FESpace}

Compute the total number of degrees of freedom.

Elfel.FEFields.setebc! — Function

setebc!(fesp::FESpace, mid, eid, comp, val::T) where {T}

Set the EBCs (essential boundary conditions).

mid = manifold dimension of the entity,

eid = serial number of the entity (term identifier),

comp = which degree of freedom in the term,

val = value of type T

For instance, mid = 0 means set the degree of freedom at the vertex eid.

Elfel.FEFields.numberfreedofs! — Function

numberfreedofs!(f::FEField, firstnum = 1)

Number the unknowns in the field, starting from the one supplied on input.

Note: The data degrees of freedom have their numbers zeroed out.

numberfreedofs!(fesp::FES, firstnum = 1)  where {FES<:FESpace}

Number the free degrees of freedom.

The unknown degrees of freedom in the FE space are numbered consecutively.

No effort is made to optimize the numbering in any way.

Elfel.FEFields.numberdatadofs! — Function

numberdatadofs!(f::FEField, firstnum = 1)

Number the data degrees of freedom in the field. Start from the number supplied on input.

Note: The free degrees of freedom are numbered first.

numberdatadofs!(fesp::FES, firstnum = 1)  where {FES<:FESpace}

Number the data (known) degrees of freedom.

The known degrees of freedom in the FE space are numbered consecutively.

No effort is made to optimize the numbering in any way.

Missing docstring.

Missing docstring for freedofnums. Check Documenter's build log for details.

Missing docstring.

Missing docstring for datadofnums. Check Documenter's build log for details.

Elfel.FEFields.highestfreedofnum — Function

highestfreedofnum(fesp::FES)  where {FES<:FESpace}

Compute the highest number of free (unknown) degrees of freedom.

Elfel.FEFields.highestdatadofnum — Function

highestfreedofnum(fesp::FES)  where {FES<:FESpace}

Compute the highest number of data (known) degrees of freedom.

Elfel.FEFields.gathersysvec! — Function

gathersysvec!(v, fesp::FESpace)

Gather values for the whole system vector.

gathersysvec!(v, fesp)

Gather values for the whole system vector from all FE spaces contributing to it.

fesp is either a vector or a tuple of FE spaces.

Elfel.FEFields.scattersysvec! — Function

scattersysvec!(fesp::FESpace, v)

Scatter values from the system vector.

scattersysvec!(fesp, v)

Scatter values for the whole system vector to all FE spaces contributing to it.

fesp is either a vector or a tuple of FE spaces.

Spaces

Elfel.FESpaces.doftype — Function

doftype(fesp::FESpace{FET, T}) where {FET, T}

Provide the type of the values of the degrees of freedom.

Missing docstring.

Missing docstring for edofmdim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for edofbfnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for edofcompnt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ndofsperel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for numberfreedofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for numberdatadofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ndofs. Check Documenter's build log for details.

Elfel.FESpaces.nunknowns — Function

nunknowns(fesp::FES)  where {FES<:FESpace}

Compute the total number of unknown degrees of freedom.

Missing docstring.

Missing docstring for highestfreedofnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for highestdatadofnum. Check Documenter's build log for details.

Elfel.FESpaces.numberdofs! — Function

numberdofs!(fesp...)

Number the degrees of freedom of a collection of FE spaces.

Missing docstring.

Missing docstring for setebc!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gathersysvec!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scattersysvec!. Check Documenter's build log for details.

Elfel.FESpaces.makeattribute — Function

makeattribute(fesp::FESpace, name, comp)

Attach attribute to the right shape collection of all incidence relations.

Finite element iterators

Missing docstring.

Missing docstring for Base.iterate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ndofsperel. Check Documenter's build log for details.

Elfel.FEIterators.eldofs — Function

eldofs(it::FEIterator)

Retrieve the vector of the element degrees of freedom

Elfel.FEIterators.elnodes — Function

elnodes(it::FEIterator)

Retrieve the vector of the nodes of the element.

Elfel.FEIterators.eldofentmdims — Function

eldofentmdims(it::FEIterator)

Retrieve the vector of the entity dimensions for each element degree of freedom.

Elfel.FEIterators.eldofcomps — Function

eldofcomps(it::FEIterator)

Retrieve the vector of the component numbers for each element degree of freedom.

Missing docstring.

Missing docstring for jacjac. Check Documenter's build log for details.

Quadrature-point iterators

Missing docstring.

Missing docstring for Base.iterate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bfun. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bfungradpar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bfungrad. Check Documenter's build log for details.

Missing docstring.

Missing docstring for weight. Check Documenter's build log for details.

Assemblers

Elfel.Assemblers.SysmatAssemblerSparse — Type

SysmatAssemblerSparse{T<:Number} <: AbstractSysmatAssembler

Type for assembling a sparse global matrix from individual entries.

Elfel.Assemblers.start! — Function

start!(self::SysmatAssemblerSparse{T}, nrow, ncol) where {T<:Number}

Start the assembly of a global matrix.

start!(self::SV,  nrow) where {SV<:AbstractSysvecAssembler, T<:Number}

Start assembly.

The method makes the buffer for the vector assembly. It must be called before the first call to the method assemble.

nrow= Total number of degrees of freedom.

start!(self::SysvecAssembler{T},
  nrow::FInt) where {T<:Number}

Start assembly.

The method makes the buffer for the vector assembly. It must be called before the first call to the method assemble.

nrow= Total number of degrees of freedom.

Elfel.Assemblers.assemble! — Function

assemble!(self::SysmatAssemblerSparse{T}, r, c, v::T) where {T<:Number}

Assemble a single entry of a rectangular matrix.

assemble!(self::SysmatAssemblerSparse{T}, lma::LocalMatrixAssembler{IT, T}) where {IT<:Integer, T<:Number}

Assemble the row numbers, column numbers, and values from a local assembler.

assemble!(self::SysmatAssemblerSparse{T}, lma::Transpose{T,LocalMatrixAssembler{IT,T}}) where {IT<:Integer, T<:Number}

Assemble the row numbers, column numbers, and values from a local assembler.

assemble!(self::SV, i, val::T) where {SV<:AbstractSysvecAssembler, T<:Number}

Assemble an elementwise vector.

The method assembles a column element vector using the vector of degree of freedom numbers for the rows.

assemble!(self::SysvecAssembler{T}, vec::MV,
  dofnums::D) where {T<:Number, MV<:AbstractArray{T}, D<:AbstractArray{FInt}}

Assemble a single value.

The method assembles a column element vector using the vector of degree of freedom numbers for the rows.

assemble!(self::SysvecAssembler{T}, rs::AbstractVector{IT}, vs::AbstractVector{T}) where {IT<:Integer, T<:Number}

Assemble entire vector.

Elfel.Assemblers.finish! — Function

finish!(self::SysmatAssemblerSparse)

Make a sparse matrix.

makevector!(self::SysvecAssembler)

Make the global vector.

finish!(self::SysvecAssembler)

Make the global vector.

Elfel.Assemblers.SysvecAssembler — Type

SysvecAssembler

Assembler for the system vector.

Local Assemblers

Missing docstring.

Missing docstring for LocalMatrixAssembler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LocalVectorAssembler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.size. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.getindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.setindex!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for init!. Check Documenter's build log for details.

Index

Elfel.Assemblers.SysmatAssemblerSparse

Elfel.Assemblers.SysvecAssembler

Elfel.Assemblers.assemble!

Elfel.Assemblers.finish!

Elfel.Assemblers.start!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.ndofs

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.setebc!

Elfel.FEIterators.eldofcomps

Elfel.FEIterators.eldofentmdims

Elfel.FEIterators.eldofs

Elfel.FEIterators.elnodes

Elfel.FESpaces.doftype

Elfel.FESpaces.makeattribute

Elfel.FESpaces.numberdofs!

Elfel.FESpaces.nunknowns

Elfel.FElements.Jacobian

Elfel.FElements.jacjac

Elfel.FElements.ndofperfeat

Elfel.FElements.ndofsperel

Elfel.FElements.nfeatofdim

Elfel.FElements.refshape

Elfel.FElements.shapedesc

Elfel.RefShapes.manifdim

Elfel.RefShapes.manifdimv

MeshCore.manifdim