Types

Reference shapes

Elfel.RefShapes.AbstractRefShape — Type

AbstractRefShape{MANIFDIM}

Abstract type of a reference shape.

Elfel.RefShapes.RefShapePoint — Type

RefShapePoint <: AbstractRefShape{0}

Type of a reference shape for a zero-dimensional manifold (point).

Elfel.RefShapes.RefShapeInterval — Type

RefShapeInterval <: AbstractRefShape{1}

Type of a reference shape for a 1-dimensional manifold (curve).

Elfel.RefShapes.RefShapeSquare — Type

RefShapeSquare <: AbstractRefShape{2}

Type of a logically rectangular reference shape for a 2-dimensional manifold (surface).

Elfel.RefShapes.RefShapeCube — Type

RefShapeCube <: AbstractRefShape{3}

Type of a reference shape for a 3-dimensional manifold (solid) bounded by six quadrilaterals.

Elfel.RefShapes.RefShapeTriangle — Type

RefShapeTriangle <: AbstractRefShape{2}

Type of a logically triangular reference shape for a 2-dimensional manifold (surface).

Elfel.RefShapes.RefShapeTetrahedron — Type

RefShapeTetrahedron <: AbstractRefShape{3}

Type of a reference shape for a 3-dimensional manifold (solid) bounded by 4 triangles.

Elfel.RefShapes.IntegRule — Type

IntegRule

Type for integration rule.

Elements

Elfel.FElements.FE — Type

FE{RS, SD}

Abstract type of finite element, parameterized by

RS: type of reference shape of the element (triangle, square, ...), and

SD: shape descriptor; refer to the package MeshCore.

Elfel.FElements.FEData — Type

FEData{SD}

Type of a finite element data.

Parameterized by

SD = shape descriptor; refer to the package MeshCore.

Fields

Elfel.FEFields.FEField — Type

FEField{N, T, IT}

Type of a finite element field. Parameterized with

N: number of degrees of freedom per entity,

T: type of the degree of freedom value,

IT: type of the index (integer value). This describes the serial numbers of the degrees of freedom.

Spaces

Elfel.FESpaces.FESpace — Type

FESpace{FET, T}

Type of a finite element space, parameterized with

FET: type of finite element, it is a scalar finite element,

T: type of degree of freedom value (double, complex, ...).

Finite element iterators

Elfel.FEIterators.FEIterator — Type

FEIterator{FES, IR, G, IT, T, V, IR0, IR1, IR2, IR3, F0, F1, F2, F3}

Type of finite element iterator. Parameterized with the types of

FES: finite element space,

IR: base incidence relation of the mesh,

G: type of the geometry attribute,

IT: type of integer indices, such as the numbers of nodes and degrees of freedom,

T: type of the degree of freedom value (real double, complex float, ... ),

V: Val representation of the manifold dimension of the base relation elements,

IR0, IR1, IR2, IR3: types of incidence relations with which degrees of freedom are associated in the finite element space, for each of the manifolds dimensions 0, 1, 2, 3,

F0, F1, F2, F3: types of fields with which degrees of freedom are associated in the finite element space, for each of the manifolds dimensions 0, 1, 2, 3.

Quadrature-point iterators

Elfel.QPIterators.QPIterator — Type

QPIterator{FES, MDIM}

Type of quadrature-point iterator, parameterized by

FES: the type of the finite element space,

MDIM: the manifold dimension of the finite element.

Assemblers

Elfel.Assemblers.AbstractSysmatAssembler — Type

AbstractSysmatAssembler

Abstract type of system-matrix assembler.

Missing docstring.

Missing docstring for SysmatAssemblerSparse{T<:Number} <: AbstractSysmatAssembler. Check Documenter's build log for details.

Elfel.Assemblers.AbstractSysvecAssembler — Type

AbstractSysvecAssembler

Abstract type of system vector assembler.

Missing docstring.

Missing docstring for SysvecAssembler{T<:Number} <: AbstractSysvecAssembler. Check Documenter's build log for details.

Local Assemblers

Missing docstring.

Missing docstring for LocalMatrixAssembler{IT<:Integer, T<:Number} <: AbstractArray{T, 2}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LocalVectorAssembler{IT<:Integer, T<:Number} <: AbstractArray{T, 1}. Check Documenter's build log for details.

Index

Elfel.Assemblers.AbstractSysmatAssembler

Elfel.Assemblers.AbstractSysvecAssembler

Elfel.Assemblers.SysmatAssemblerSparse

Elfel.Assemblers.SysvecAssembler

Elfel.FEFields.FEField

Elfel.FEIterators.FEIterator

Elfel.FESpaces.FESpace

Elfel.FElements.FE

Elfel.FElements.FEData

Elfel.LocalAssemblers.LocalMatrixAssembler

Elfel.LocalAssemblers.LocalVectorAssembler

Elfel.QPIterators.QPIterator

Elfel.RefShapes.AbstractRefShape

Elfel.RefShapes.IntegRule

Elfel.RefShapes.RefShapeCube

Elfel.RefShapes.RefShapeInterval

Elfel.RefShapes.RefShapePoint

Elfel.RefShapes.RefShapeSquare

Elfel.RefShapes.RefShapeTetrahedron

Elfel.RefShapes.RefShapeTriangle

Base.getindex

Base.iterate

Base.setindex!

Base.size

Elfel.Assemblers.assemble!

Elfel.Assemblers.finish!

Elfel.Assemblers.start!

Elfel.FEFields.datadofnums

Elfel.FEFields.dofnumtype

Elfel.FEFields.doftype

Elfel.FEFields.freedofnums

Elfel.FEFields.gathersysvec!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.ndofs

Elfel.FEFields.ndofs

Elfel.FEFields.ndofsperterm

Elfel.FEFields.nterms

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.setebc!

Elfel.FEFields.setebc!

Elfel.FEIterators.eldofcomps

Elfel.FEIterators.eldofentmdims

Elfel.FEIterators.eldofs

Elfel.FEIterators.elnodes

Elfel.FESpaces.doftype

Elfel.FESpaces.edofbfnum

Elfel.FESpaces.edofcompnt

Elfel.FESpaces.edofmdim

Elfel.FESpaces.makeattribute

Elfel.FESpaces.numberdofs!

Elfel.FESpaces.nunknowns

Elfel.FElements.FEH1_L2

Elfel.FElements.FEH1_Q4

Elfel.FElements.FEH1_T3

Elfel.FElements.FEH1_T3_BUBBLE

Elfel.FElements.FEH1_T6

Elfel.FElements.Jacobian

Elfel.FElements.bfun

Elfel.FElements.bfungradpar

Elfel.FElements.jacjac

Elfel.FElements.jacjac

Elfel.FElements.ndofperfeat

Elfel.FElements.ndofsperel

Elfel.FElements.ndofsperel

Elfel.FElements.ndofsperel

Elfel.FElements.nfeatofdim

Elfel.FElements.refshape

Elfel.FElements.shapedesc

Elfel.LocalAssemblers.init!

Elfel.QPIterators.bfungrad

Elfel.QPIterators.weight

Elfel.RefShapes.manifdim

Elfel.RefShapes.manifdimv

Elfel.RefShapes.quadrature

MeshCore.manifdim