Functions

Reference shapes

Elfel.RefShapes.manifdim — Function

manifdim(rs)

Get the manifold dimension of the reference shape.

Elfel.RefShapes.manifdimv — Function

manifdimv(::Type{T}) where {T<:AbstractRefShape{MANIFDIM}} where {MANIFDIM}

Get the manifold dimension of the reference shape as Val.

Missing docstring.

Missing docstring for quadrature. Check Documenter's build log for details.

Elements

Elfel.FElements.shapedesc — Function

shapedesc(fe::FE{RS, SD}) where {RS, SD}

Topological shape description.

Refer to the MeshCore library.

Elfel.FElements.refshape — Function

refshape(fe::FE{RS, SD}) where {RS, SD}

Reference shape.

Elfel.FElements.nfeatofdim — Function

nfeatofdim(fe::FE{RS, SD}, m) where {RS, SD}

Number of features of manifold dimension m. Note that 0 <= m <= 3.

Elfel.FElements.ndofperfeat — Function

feathasdof(fe::FE{RS, SD}, m) where {RS, SD}

How many degrees of freedom are attached to a the feature of manifold dimension m?

Note that 0 <= m <= 3.

Elfel.FElements.ndofsperel — Function

ndofsperel(it::FEIterator)

Retrieve the number of degrees of freedom per element.

ndofsperel(fe::FE{RS, SD}) where {RS, SD}

Provide the number of degrees of freedom per element.

Enumerate all features of all manifold dimensions, and for each feature multiply by the number of degrees of freedom per feature. The assumption is that this is a scalar finite element.

MeshCore.manifdim — Function

manifdim(fe::FE{RS, SD}) where {RS, SD}

Get the manifold dimension of the finite element.

Elfel.FElements.Jacobian — Function

Jacobian(::Val{0}, J::T) where {T}

Evaluate the point Jacobian.

J = Jacobian matrix, which isn't really defined well for a 0-manifold.

Jacobian(::Val{1}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Jacobian(::Val{2}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Jacobian(self::T, J::FFltMat)::FFlt where {T<:AbstractFESet3Manifold}

Evaluate the volume Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Elfel.FElements.jacjac — Function

jacjac(it::QPIterator)

Compute the Jacobian matrix and the Jacobian determinant.

At the current integration point.

jacjac(fe::FE{RS, SD}, locs, gradNpar) where {RS, SD}

Compute the Jacobian matrix and the Jacobian determinant.

This is the generic version suitable for isoparametric elements.

Missing docstring.

Missing docstring for FEH1_L2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_Q4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T6. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T3_BUBBLE. Check Documenter's build log for details.

Fields

Spaces

Finite element iterators

Quadrature-point iterators

Assemblers

Local Assemblers

Index

Elfel.FElements.Jacobian

Elfel.FElements.jacjac

Elfel.FElements.ndofperfeat

Elfel.FElements.ndofsperel

Elfel.FElements.nfeatofdim

Elfel.FElements.refshape

Elfel.FElements.shapedesc

Elfel.RefShapes.manifdim

Elfel.RefShapes.manifdimv

MeshCore.manifdim