Functions

Reference shapes

Missing docstring.

Missing docstring for Elfel.RefShapes.manifdim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.RefShapes.manifdimv. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.RefShapes.quadrature. Check Documenter's build log for details.

Elements

Missing docstring.

Missing docstring for Elfel.FElements.shapedesc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.refshape. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.nfeatofdim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.ndofperfeat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.ndofsperel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.manifdim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.Jacobian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.jacjac. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.bfun. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.bfungradpar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.FEH1_L2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.FEH1_T3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.FEH1_Q4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.FEH1_T6. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FElements.FEH1_T3_BUBBLE. Check Documenter's build log for details.

Fields

Missing docstring.

Missing docstring for Elfel.FEFields.doftype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.dofnumtype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.nterms. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.ndofsperterm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.ndofs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.setebc!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.numberfreedofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.numberdatadofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.freedofnums. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.datadofnums. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.highestfreedofnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.highestdatadofnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.gathersysvec!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEFields.scattersysvec!. Check Documenter's build log for details.

Spaces

Missing docstring.

Missing docstring for Elfel.FESpaces.doftype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.edofmdim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.edofbfnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.edofcompnt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.ndofsperel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.numberfreedofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.numberdatadofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.ndofs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.nunknowns. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.highestfreedofnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.highestdatadofnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.numberdofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.setebc!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.gathersysvec!(v, fesp::FESpace). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.gathersysvec!(v, fesp::AbstractVector). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.scattersysvec!(fesp::FESpace, v). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.scattersysvec!(fesp::AbstractVector, v). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FESpaces.makeattribute. Check Documenter's build log for details.

Finite element iterators

Base.iterate — Function

Base.iterate(it::QPIterator, state = 1)

Advance a quadrature point iterator.

Base.iterate(it::FEIterator, state = 1)

Advance the iterator to the next entity.

The nodes of the finite element are cached, as is a vector of all the degrees of freedom represented on the element.

Missing docstring.

Missing docstring for Elfel.FEIterators.ndofsperel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEIterators.eldofs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEIterators.elnodes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEIterators.eldofentmdims. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEIterators.eldofcomps. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.FEIterators.jacjac. Check Documenter's build log for details.

Quadrature-point iterators

Missing docstring.

Missing docstring for Base.iterate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.QPIterators.bfun. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.QPIterators.bfungradpar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.QPIterators.bfungrad. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.QPIterators.weight. Check Documenter's build log for details.

Assemblers

Missing docstring.

Missing docstring for Elfel.Assemblers.SysmatAssemblerSparse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.Assemblers.start!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.Assemblers.assemble!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.Assemblers.finish!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.Assemblers.SysvecAssembler. Check Documenter's build log for details.

Local Assemblers

Missing docstring.

Missing docstring for Elfel.LocalAssemblers.LocalMatrixAssembler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Elfel.LocalAssemblers.LocalVectorAssembler. Check Documenter's build log for details.

Base.size — Function

Base.size(a::A) where {A<:LocalMatrixAssembler}

The size is the tuple of number of rows and number of columns.

Base.size(a::A) where {A<:LocalVectorAssembler}

The size is the number of rows (entries).

Base.getindex — Function

Base.getindex(a::A, i::Int, j::Int)

Only access to a single entry of the matrix is provided.

Base.getindex(a::A, i::Int) where {A<:LocalVectorAssembler}

Access is provided to a single entry of the vector.

Base.setindex! — Function

Base.setindex!(a::A, v, i::Int, j::Int) where {A<:LocalMatrixAssembler}

Only access to a single entry of the matrix is provided.

Base.setindex!(a::A, v, i::Int) where {A<:LocalVectorAssembler}

Access is provided to a single entry of the vector.

Missing docstring.

Missing docstring for Elfel.LocalAssemblers.init!. Check Documenter's build log for details.

Index

Elfel.Assemblers.AbstractSysmatAssembler

Elfel.Assemblers.AbstractSysvecAssembler

Elfel.Assemblers.SysmatAssemblerSparse

Elfel.Assemblers.SysvecAssembler

Elfel.FEFields.FEField

Elfel.FEIterators.FEIterator

Elfel.FESpaces.FESpace

Elfel.FElements.FE

Elfel.FElements.FEData

Elfel.LocalAssemblers.LocalMatrixAssembler

Elfel.LocalAssemblers.LocalVectorAssembler

Elfel.QPIterators.QPIterator

Elfel.RefShapes.AbstractRefShape

Elfel.RefShapes.IntegRule

Elfel.RefShapes.RefShapeCube

Elfel.RefShapes.RefShapeInterval

Elfel.RefShapes.RefShapePoint

Elfel.RefShapes.RefShapeSquare

Elfel.RefShapes.RefShapeTetrahedron

Elfel.RefShapes.RefShapeTriangle

Base.getindex

Base.iterate

Base.setindex!

Base.size