Functions

Reference shapes

Elfel.RefShapes.manifdim — Function

manifdim(rs)

Get the manifold dimension of the reference shape.

Elfel.RefShapes.manifdimv — Function

manifdimv(::Type{T}) where {T<:AbstractRefShape{MANIFDIM}} where {MANIFDIM}

Get the manifold dimension of the reference shape as Val.

Elfel.RefShapes.quadrature — Function

quadrature(::Type{RefShapeInterval}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of an interval.

The default is Gauss integration rule, where the order is set with the keyword order.

quadrature(::Type{RefShapeTriangle}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of an triangle.

The default is a triangle rule, distinguished by the number of points set with the keyword npts.

quadrature(::Type{RefShapeSquare}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of a square.

The default is Gauss integration rule, where the order is set with the keyword order.

Elements

Elfel.FElements.shapedesc — Function

shapedesc(fe::FE{RS, SD}) where {RS, SD}

Topological shape description.

Refer to the MeshCore library.

Elfel.FElements.refshape — Function

refshape(fe::FE{RS, SD}) where {RS, SD}

Reference shape.

Elfel.FElements.nfeatofdim — Function

nfeatofdim(fe::FE{RS, SD}, m) where {RS, SD}

Number of features of manifold dimension m. Note that 0 <= m <= 3.

Elfel.FElements.ndofperfeat — Function

feathasdof(fe::FE{RS, SD}, m) where {RS, SD}

How many degrees of freedom are attached to a the feature of manifold dimension m?

Note that 0 <= m <= 3.

Elfel.FElements.ndofsperel — Function

ndofsperel(fe::FE{RS, SD}) where {RS, SD}

Provide the number of degrees of freedom per element.

Enumerate all features of all manifold dimensions, and for each feature multiply by the number of degrees of freedom per feature. The assumption is that this is a scalar finite element.

ndofsperel(fesp::FES)  where {FES<:FESpace}

Total number of degrees of freedom associated with each finite element.

Essentially a product of the number of the degrees of freedom the scalar finite element and the number of copies of this element in the space.

ndofsperel(it::FEIterator)

Retrieve the number of degrees of freedom per element.

MeshCore.manifdim — Function

manifdim(fe::FE{RS, SD}) where {RS, SD}

Get the manifold dimension of the finite element.

Elfel.FElements.Jacobian — Function

Jacobian(::Val{0}, J::T) where {T}

Evaluate the point Jacobian.

J = Jacobian matrix, which isn't really defined well for a 0-manifold.

Jacobian(::Val{1}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Jacobian(::Val{2}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Jacobian(self::T, J::FFltMat)::FFlt where {T<:AbstractFESet3Manifold}

Evaluate the volume Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Elfel.FElements.jacjac — Function

jacjac(fe::FE{RS, SD}, locs, gradNpar) where {RS, SD}

Compute the Jacobian matrix and the Jacobian determinant.

This is the generic version suitable for isoparametric elements.

jacjac(it::FEIterator, qpit::QPIterator)

Compute the Jacobian matrix and the Jacobian determinant.

The finite element iterator cooperates with the quadrature point iterator here to compute the Jacobian at the current integration point.

Missing docstring.

Missing docstring for FEH1_L2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_Q4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T6. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FEH1_T3_BUBBLE. Check Documenter's build log for details.

Fields

Missing docstring.

Missing docstring for doftype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dofnumtype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nterms. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ndofsperterm. Check Documenter's build log for details.

Elfel.FEFields.ndofs — Function

ndofs(fesp::FES)  where {FES<:FESpace}

Compute the total number of degrees of freedom.

Elfel.FEFields.setebc! — Function

setebc!(fesp::FESpace, mid, eid, comp, val::T) where {T}

Set the EBCs (essential boundary conditions).

mid = manifold dimension of the entity,

eid = serial number of the entity (term identifier),

comp = which degree of freedom in the term,

val = value of type T

For instance, mid = 0 means set the degree of freedom at the vertex eid.

Elfel.FEFields.numberfreedofs! — Function

numberfreedofs!(fesp::FES, firstnum = 1)  where {FES<:FESpace}

Number the free degrees of freedom.

The unknown degrees of freedom in the FE space are numbered consecutively.

No effort is made to optimize the numbering in any way.

numberfreedofs!(f::FEField, firstnum = 1)

Number the unknowns in the field, starting from the one supplied on input.

Note: The data degrees of freedom have their numbers zeroed out.

Elfel.FEFields.numberdatadofs! — Function

numberdatadofs!(fesp::FES, firstnum = 1)  where {FES<:FESpace}

Number the data (known) degrees of freedom.

The known degrees of freedom in the FE space are numbered consecutively.

No effort is made to optimize the numbering in any way.

numberdatadofs!(f::FEField, firstnum = 1)

Number the data degrees of freedom in the field. Start from the number supplied on input.

Note: The free degrees of freedom are numbered first.

Missing docstring.

Missing docstring for freedofnums. Check Documenter's build log for details.

Missing docstring.

Missing docstring for datadofnums. Check Documenter's build log for details.

Elfel.FEFields.highestfreedofnum — Function

highestfreedofnum(fesp::FES)  where {FES<:FESpace}

Compute the highest number of free (unknown) degrees of freedom.

Elfel.FEFields.highestdatadofnum — Function

highestfreedofnum(fesp::FES)  where {FES<:FESpace}

Compute the highest number of data (known) degrees of freedom.

Elfel.FEFields.gathersysvec! — Function

gathersysvec!(v, fesp::FESpace)

Gather values for the whole system vector.

gathersysvec!(v, fesp)

Gather values for the whole system vector from all FE spaces contributing to it.

fesp is either a vector or a tuple of FE spaces.

Elfel.FEFields.scattersysvec! — Function

scattersysvec!(fesp::FESpace, v)

Scatter values from the system vector.

scattersysvec!(fesp, v)

Scatter values for the whole system vector to all FE spaces contributing to it.

fesp is either a vector or a tuple of FE spaces.

Spaces

Elfel.FESpaces.doftype — Function

doftype(fesp::FESpace{FET, T}) where {FET, T}

Provide the type of the values of the degrees of freedom.

Elfel.FESpaces.edofmdim — Function

edofmdim(fesp::FESpace{FET, T}) where {FET, T}

Access vector of manifold dimensions of entities associated with each degree of freedom.

Elfel.FESpaces.edofbfnum — Function

edofbfnum(fesp::FESpace{FET, T}) where {FET, T}

Access vector of numbers of basis functions associated with each degree of freedom.

Elfel.FESpaces.edofcompnt — Function

edofcompnt(fesp::FESpace{FET, T}) where {FET, T}

Access vector of component number associated with each degree of freedom.

When the finite element space consists of multiple copies of the scalar finite element, the component is the serial number of the copy.

Missing docstring.

Missing docstring for ndofsperel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for numberfreedofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for numberdatadofs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ndofs. Check Documenter's build log for details.

Elfel.FESpaces.nunknowns — Function

nunknowns(fesp::FES)  where {FES<:FESpace}

Compute the total number of unknown degrees of freedom.

Missing docstring.

Missing docstring for highestfreedofnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for highestdatadofnum. Check Documenter's build log for details.

Elfel.FESpaces.numberdofs! — Function

numberdofs!(fesp...)

Number the degrees of freedom of a collection of FE spaces.

The unknown (free) degrees of freedom in the FE space are numbered consecutively, and then the data degrees of freedom (the known values) are numbered.

No effort is made to optimize the numbering in any way.

Missing docstring.

Missing docstring for setebc!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gathersysvec!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scattersysvec!. Check Documenter's build log for details.

Elfel.FESpaces.makeattribute — Function

makeattribute(fesp::FESpace, name, comp)

Attach attribute to the right shape collection of all incidence relations.

Finite element iterators

Base.iterate — Function

Base.iterate(it::FEIterator, state = 1)

Advance the iterator to the next entity.

The nodes of the finite element are cached, as is a vector of all the degrees of freedom represented on the element.

Missing docstring.

Missing docstring for ndofsperel. Check Documenter's build log for details.

Elfel.FEIterators.eldofs — Function

eldofs(it::FEIterator)

Retrieve the vector of the element degrees of freedom

Elfel.FEIterators.elnodes — Function

elnodes(it::FEIterator)

Retrieve the vector of the nodes of the element.

Elfel.FEIterators.eldofentmdims — Function

eldofentmdims(it::FEIterator)

Retrieve the vector of the entity dimensions for each element degree of freedom.

Each degree of freedom is associated with some entity of the finite element: vertices, edges, faces, and so on. This vector records the dimension of the manifold entity with which each degree of freedom is associated.

Elfel.FEIterators.eldofcomps — Function

eldofcomps(it::FEIterator)

Retrieve the vector of the component numbers for each element degree of freedom.

If multiple copies of the finite element are referenced in the finite element space, each copy is referred to as component.

Missing docstring.

Missing docstring for jacjac. Check Documenter's build log for details.

Quadrature-point iterators

Missing docstring.

Missing docstring for Base.iterate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bfun. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bfungradpar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bfungrad. Check Documenter's build log for details.

Missing docstring.

Missing docstring for weight. Check Documenter's build log for details.

Assemblers

Elfel.Assemblers.SysmatAssemblerSparse — Type

SysmatAssemblerSparse{T<:Number} <: AbstractSysmatAssembler

Type for assembling a sparse global matrix from individual entries.

Elfel.Assemblers.start! — Function

start!(self::SysmatAssemblerSparse{T}, nrow, ncol) where {T<:Number}

Start the assembly of a global matrix.

start!(self::SV,  nrow) where {SV<:AbstractSysvecAssembler, T<:Number}

Start assembly.

The method makes the buffer for the vector assembly. It must be called before the first call to the method assemble.

nrow= Total number of degrees of freedom.

start!(self::SysvecAssembler{T},  nrow::Int64) where {T<:Number}

Start assembly.

The method makes the buffer for the vector assembly. It must be called before the first call to the method assemble.

nrow= Total number of degrees of freedom.

Elfel.Assemblers.assemble! — Function

assemble!(self::SysmatAssemblerSparse{T}, r, c, v::T) where {T<:Number}

Assemble a single entry of a rectangular matrix.

assemble!(self::SysmatAssemblerSparse{T}, lma::LocalMatrixAssembler{IT, T}) where {IT<:Integer, T<:Number}

Assemble the row numbers, column numbers, and values from a local assembler.

assemble!(self::SysmatAssemblerSparse{T}, lma::Transpose{T,LocalMatrixAssembler{IT,T}}) where {IT<:Integer, T<:Number}

Assemble the row numbers, column numbers, and values from a local assembler.

assemble!(self::SV, i, val::T) where {SV<:AbstractSysvecAssembler, T<:Number}

Assemble an elementwise vector.

The method assembles a column element vector using the vector of degree of freedom numbers for the rows.

assemble!(self::SysvecAssembler{T}, i, val::T) where {T<:Number}

Assemble a single value into the row i.

assemble!(self::SysvecAssembler{T}, lva) where {T<:Number}

Assemble local vector assembler.

Elfel.Assemblers.finish! — Function

finish!(self::SysmatAssemblerSparse)

Make a sparse matrix.

makevector!(self::SysvecAssembler)

Make the global vector.

finish!(self::SysvecAssembler)

Make the global vector.

Elfel.Assemblers.SysvecAssembler — Type

SysvecAssembler

Assembler for the system vector.

Local Assemblers

Elfel.LocalAssemblers.LocalMatrixAssembler — Type

LocalMatrixAssembler{IT<:Integer, T<:Number} <: AbstractArray{T, 2}

Type of "local" matrix assembler.

Local is to be understood in the sense of in the context of a single finite element. So a local matrix is an elementwise matrix which is computed entry by entry. Then it can be assembled into the global matrix in one shot.

Elfel.LocalAssemblers.LocalVectorAssembler — Type

LocalVectorAssembler{IT<:Integer, T<:Number} <: AbstractArray{T, 1}

Type of "local" vector assembler.

Local is to be understood in the sense of in the context of a single finite element. So a local vector is an elementwise vector which is computed entry by entry. Then it can be assembled into the global vector in one shot.

Base.size — Function

Base.size(a::A) where {A<:LocalMatrixAssembler}

The size is the tuple of number of rows and number of columns.

Base.size(a::A) where {A<:LocalVectorAssembler}

The size is the number of rows (entries).

Base.getindex — Function

Base.getindex(a::A, i::Int, j::Int)

Only access to a single entry of the matrix is provided.

Base.getindex(a::A, i::Int) where {A<:LocalVectorAssembler}

Access is provided to a single entry of the vector.

Base.setindex! — Function

Base.setindex!(a::A, v, i::Int, j::Int) where {A<:LocalMatrixAssembler}

Only access to a single entry of the matrix is provided.

Base.setindex!(a::A, v, i::Int) where {A<:LocalVectorAssembler}

Access is provided to a single entry of the vector.

Elfel.LocalAssemblers.init! — Function

init!(a::L, rdofs, cdofs) where {L<:LocalMatrixAssembler{IT, T}} where {IT, T}

Initialize the local assembler with the global degrees of freedom in the rows and columns.

The two arrays, rdofs, cdofs, define the global degree of freedom numbers for the element. The data matrix is zeroed out.

This function needs to be called for each new finite element.

init!(a::L, rdofs) where {L<:LocalVectorAssembler{IT, T}} where {IT, T}

Initialize the local assembler with the global degrees of freedom in the rows.

The array rdofs defines the global degree of freedom numbers for the element. The data vector is zeroed out.

This function needs to be called for each new finite element.

Index

Elfel.Assemblers.SysmatAssemblerSparse

Elfel.Assemblers.SysvecAssembler

Elfel.LocalAssemblers.LocalMatrixAssembler

Elfel.LocalAssemblers.LocalVectorAssembler

Base.getindex

Base.iterate

Base.setindex!

Base.size

Elfel.Assemblers.assemble!

Elfel.Assemblers.finish!

Elfel.Assemblers.start!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.ndofs

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.setebc!

Elfel.FEIterators.eldofcomps

Elfel.FEIterators.eldofentmdims

Elfel.FEIterators.eldofs

Elfel.FEIterators.elnodes

Elfel.FESpaces.doftype

Elfel.FESpaces.edofbfnum

Elfel.FESpaces.edofcompnt

Elfel.FESpaces.edofmdim

Elfel.FESpaces.makeattribute

Elfel.FESpaces.numberdofs!

Elfel.FESpaces.nunknowns

Elfel.FElements.Jacobian

Elfel.FElements.jacjac

Elfel.FElements.ndofperfeat

Elfel.FElements.ndofsperel

Elfel.FElements.nfeatofdim

Elfel.FElements.refshape

Elfel.FElements.shapedesc

Elfel.LocalAssemblers.init!

Elfel.RefShapes.manifdim

Elfel.RefShapes.manifdimv

Elfel.RefShapes.quadrature

MeshCore.manifdim