Reference manual

Reference shapes

Elfel.RefShapes.AbstractRefShape — Type

AbstractRefShape{MANIFDIM}

Abstract type of a reference shape.

Elfel.RefShapes.IntegRule — Type

IntegRule

Type for integration rule.

Elfel.RefShapes.RefShapeCube — Type

RefShapeCube <: AbstractRefShape{3}

Type of a reference shape for a 3-dimensional manifold (solid) bounded by six quadrilaterals.

Elfel.RefShapes.RefShapeInterval — Type

RefShapeInterval <: AbstractRefShape{1}

Type of a reference shape for a 1-dimensional manifold (curve).

Elfel.RefShapes.RefShapePoint — Type

RefShapePoint <: AbstractRefShape{0}

Type of a reference shape for a zero-dimensional manifold (point).

Elfel.RefShapes.RefShapeSquare — Type

RefShapeSquare <: AbstractRefShape{2}

Type of a logically rectangular reference shape for a 2-dimensional manifold (surface).

Elfel.RefShapes.RefShapeTetrahedron — Type

RefShapeTetrahedron <: AbstractRefShape{3}

Type of a reference shape for a 3-dimensional manifold (solid) bounded by 4 triangles.

Elfel.RefShapes.RefShapeTriangle — Type

RefShapeTriangle <: AbstractRefShape{2}

Type of a logically triangular reference shape for a 2-dimensional manifold (surface).

Elfel.RefShapes.manifdim — Method

manifdim(rs)

Get the manifold dimension of the reference shape.

Elfel.RefShapes.manifdimv — Method

manifdimv(::Type{T}) where {T<:AbstractRefShape{MANIFDIM}} where {MANIFDIM}

Get the manifold dimension of the reference shape as Val.

Elfel.RefShapes.quadrature — Function

quadrature(::Type{RefShapeTetrahedron}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of a tetrahedron.

The default is a one-point tetrahedron rule; other rules may be chosen based on the number of points set with the keyword npts.

Elfel.RefShapes.quadrature — Function

quadrature(::Type{RefShapeSquare}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of a square.

The default is Gauss integration rule, where the order is set with the keyword order.

Elfel.RefShapes.quadrature — Function

quadrature(::Type{RefShapeTriangle}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of an triangle.

The default is a triangle rule, distinguished by the number of points set with the keyword npts.

Elfel.RefShapes.quadrature — Function

quadrature(::Type{RefShapeInterval}, quadraturesettings = (kind = :default,))

Create a quadrature rule for the reference shape of an interval.

The default is Gauss integration rule, where the order is set with the keyword order.

Elements

Elfel.FElements.FE — Type

FE{RS, SD}

Abstract type of finite element, parameterized by

RS: type of reference shape of the element (triangle, square, ...), and

SD: shape descriptor; refer to the package MeshCore.

Elfel.FElements.FEData — Type

FEData{SD}

Type of a finite element data.

Parameterized by

SD = shape descriptor; refer to the package MeshCore.

Elfel.FElements.FEH1_L2 — Method

FEH1_L2()

Construct an H1 finite element of the type L2.

L2 is two-node linear segment element.

Elfel.FElements.FEH1_Q4 — Method

FEH1_Q4()

Construct an H1 finite element of the type Q4.

Q4 is 4-node linear quadrilateral element.

Elfel.FElements.FEH1_T3 — Method

FEH1_T3()

Construct an H1 finite element of the type T3.

T3 is 3-node linear triangle element.

Elfel.FElements.FEH1_T3_BUBBLE — Method

FEH1_T3_BUBBLE()

Construct an H1 finite element of the type T3 with a cubic bubble.

T3 is 3-node linear triangle element with a cubic bubble. It has the usual nodal basis functions associated with the vertices, and cubic bubble associated with the element itself.

Elfel.FElements.FEH1_T4 — Method

FEH1_T4()

Construct an H1 finite element of the type T4.

T4 is 4-node linear tetrahedral element.

Elfel.FElements.FEH1_T6 — Method

FEH1_T6()

Construct an H1 finite element of the type T6.

T6 is 6-node quadratic triangle element.

Elfel.FElements.FEL2_Q4 — Method

FEL2_Q4()

Construct an L2 finite element of the type Q4.

Q4 is 4-node linear quadrilateral element.

To do: Enable multiple basis functions in the interior of the element; perhaps linear.

Elfel.FElements.Jacobian — Method

Jacobian(::Val{0}, J::T) where {T}

Evaluate the point Jacobian.

J = Jacobian matrix, which isn't really defined well for a 0-manifold.

Elfel.FElements.Jacobian — Method

Jacobian(::Val{1}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Elfel.FElements.Jacobian — Method

Jacobian(::Val{2}, J::T) where {T}

Evaluate the curve Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Elfel.FElements.Jacobian — Method

Jacobian(self::T, J::FFltMat)::FFlt where {T<:AbstractFESet3Manifold}

Evaluate the volume Jacobian.

J = Jacobian matrix, columns are tangent to parametric coordinates curves.

Elfel.FElements.bfun — Method

bfun(self::FESUBT,  param_coords)  where {FESUBT<:FE{RS, SD}}

Evaluate the basis functions for all degrees of freedom of the scalar finite element at the parametric coordinates. Return a vector of the values.

Elfel.FElements.bfungradpar — Method

bfungradpar(self::FESUBT,  param_coords)  where {FESUBT<:FE{RS, SD}}

Evaluate the gradients of the basis functions for all degrees of freedom of the scalar finite element with respect to the parametric coordinates, at the parametric coordinates given. Return a vector of the gradients.

Elfel.FElements.jacjac — Method

jacjac(fe::FE{RS, SD}, locs, nodes, gradNpar) where {RS, SD}

Compute the Jacobian matrix and the Jacobian determinant.

This is the generic version suitable for isoparametric elements.

Elfel.FElements.ndofperfeat — Method

feathasdof(fe::FE{RS, SD}, m) where {RS, SD}

How many degrees of freedom are attached to a the feature of manifold dimension m?

Note that 0 <= m <= 3.

Elfel.FElements.ndofsperel — Method

ndofsperel(fe::FE{RS, SD}) where {RS, SD}

Provide the number of degrees of freedom per element.

Enumerate all features of all manifold dimensions, and for each feature multiply by the number of degrees of freedom per feature. The assumption is that this is a scalar finite element.

Elfel.FElements.nfeatofdim — Method

nfeatofdim(fe::FE{RS, SD}, m) where {RS, SD}

Number of features of manifold dimension m. Note that 0 <= m <= 3.

Elfel.FElements.refshape — Method

refshape(fe::FE{RS, SD}) where {RS, SD}

Reference shape.

Elfel.FElements.shapedesc — Method

shapedesc(fe::FE{RS, SD}) where {RS, SD}

Topological shape description.

Refer to the MeshCore library.

MeshCore.manifdim — Method

manifdim(fe::FE{RS, SD}) where {RS, SD}

Get the manifold dimension of the finite element.

Fields

Elfel.FEFields.FEField — Type

FEField{N, T, IT}

Type of a finite element field. Parameterized with

N: number of degrees of freedom per entity,

T: type of the degree of freedom value,

IT: type of the index (integer value). This describes the serial numbers of the degrees of freedom.

Elfel.FEFields.datadofnums — Method

datadofnums(f::FEField)

Collect information about known (data) degree of freedom numbers.

First number, last number, and the total number of degrees of freedom are returned as a tuple.

Elfel.FEFields.dofnumtype — Method

dofnumtype(fef::FEField{N, T, IT}) where {N, T, IT}

Type of the index (serial number) of the degree of freedom. Integer.

Elfel.FEFields.doftype — Method

doftype(fef::FEField{N, T, IT}) where {N, T, IT}

Type of a degree of freedom value.

Elfel.FEFields.freedofnums — Method

freedofnums(f::FEField)

Collect information about unknown (free) degree of freedom numbers.

First number, last number, and the total number of degrees of freedom are returned as a tuple.

Elfel.FEFields.gathersysvec! — Method

gathersysvec!(vec, self::FEField)

Gather system vector contributions from the field.

Elfel.FEFields.highestdatadofnum — Method

highestdatadofnum(f::FEField)

Compute the highest serial number of a datum degree of freedom in the field.

Elfel.FEFields.highestfreedofnum — Method

highestfreedofnum(f::FEField)

Compute the highest serial number of a free degree of freedom in the field.

Elfel.FEFields.ndofs — Method

ndofs(fef::FEField)

Total number of degrees of freedom in the field.

Elfel.FEFields.ndofsperterm — Method

ndofsperterm(fef::FEField{N}) where {N}

Number of degrees of freedom per term of the field.

Elfel.FEFields.nterms — Method

nterms(fef::FEField)

Number of terms in the field.

Elfel.FEFields.numberdatadofs! — Function

numberdatadofs!(f::FEField, firstnum = 1)

Number the data degrees of freedom in the field. Start from the number supplied on input.

Note: The free degrees of freedom must be numbered first.

Elfel.FEFields.numberfreedofs! — Function

numberfreedofs!(f::FEField, firstnum = 1)

Number the unknowns in the field, starting from the one supplied on input.

Note: The data degrees of freedom have their numbers zeroed out.

Elfel.FEFields.scattersysvec! — Method

scattersysvec!(self::FEField, v)

Scatter a system vector into the field.

Elfel.FEFields.setebc! — Method

setebc!(self::FEField, tid, comp, val::T) where {T}

Set the value of one particular degree of freedom to a given number.

tid: which term,

comp: which component of the term,

val: value to which the degree of freedom should be set.

Spaces

Elfel.FESpaces.FESpace — Type

FESpace{FET, T}

Type of a finite element space, parameterized with

FET: type of finite element, it is a scalar finite element,

T: type of degree of freedom value (double, complex, ...).

Elfel.FEFields.highestdatadofnum — Method

highestfreedofnum(fesp::FES)  where {FES<:FESpace}

Compute the highest number of data (known) degrees of freedom.

Elfel.FEFields.highestfreedofnum — Method

highestfreedofnum(fesp::FES)  where {FES<:FESpace}

Compute the highest number of free (unknown) degrees of freedom.

Elfel.FEFields.ndofs — Method

ndofs(fesp::FES)  where {FES<:FESpace}

Compute the total number of degrees of freedom.

Elfel.FEFields.numberdatadofs! — Method

numberdatadofs!(fesp::FES, firstnum = 1)  where {FES<:FESpace}

Number the data (known) degrees of freedom.

The known degrees of freedom in the FE space are numbered consecutively.

No effort is made to optimize the numbering in any way.

Elfel.FEFields.numberfreedofs! — Method

numberfreedofs!(fesp::FES, firstnum = 1)  where {FES<:FESpace}

Number the free degrees of freedom.

The unknown degrees of freedom in the FE space are numbered consecutively.

No effort is made to optimize the numbering in any way.

Elfel.FEFields.setebc! — Method

setebc!(fesp::FESpace, mid, eid, comp, val::T) where {T}

Set the EBCs (essential boundary conditions).

mid = manifold dimension of the entity,

eid = serial number of the entity (term identifier),

comp = which degree of freedom in the term,

val = value of type T

For instance, mid = 0 means set the degree of freedom at the vertex eid.

Elfel.FESpaces.doftype — Method

doftype(fesp::FESpace{FET, T}) where {FET, T}

Provide the type of the values of the degrees of freedom.

Elfel.FESpaces.edofbfnum — Method

edofbfnum(fesp::FESpace{FET, T}) where {FET, T}

Access vector of numbers of basis functions associated with each degree of freedom.

Elfel.FESpaces.edofcompnt — Method

edofcompnt(fesp::FESpace{FET, T}) where {FET, T}

Access vector of component number associated with each degree of freedom.

When the finite element space consists of multiple copies of the scalar finite element, the component is the serial number of the copy.

Elfel.FESpaces.edofmdim — Method

edofmdim(fesp::FESpace{FET, T}) where {FET, T}

Access vector of manifold dimensions of entities associated with each degree of freedom.

Elfel.FESpaces.makeattribute — Method

makeattribute(fesp::FESpace, name, comp)

Attach attribute to the right shape collection of all incidence relations.

Elfel.FESpaces.numberdofs! — Method

numberdofs!(fesp...)

Number the degrees of freedom of a collection of FE spaces.

The unknown (free) degrees of freedom in the FE space are numbered consecutively, and then the data degrees of freedom (the known values) are numbered.

No effort is made to optimize the numbering in any way.

Elfel.FESpaces.nunknowns — Method

nunknowns(fesp::FES)  where {FES<:FESpace}

Compute the total number of unknown degrees of freedom.

Elfel.FElements.ndofsperel — Method

ndofsperel(fesp::FES)  where {FES<:FESpace}

Total number of degrees of freedom associated with each finite element.

Essentially a product of the number of the degrees of freedom the scalar finite element and the number of copies of this element in the space.

Finite element iterators

Elfel.FEIterators.FEIterator — Type

FEIterator{FES, IR, G, IT, T, V, IR0, IR1, IR2, IR3, F0, F1, F2, F3}

Type of finite element iterator. Parameterized with the types of

FES: finite element space,

IR: base incidence relation of the mesh,

G: type of the geometry attribute,

IT: type of integer indices, such as the numbers of nodes and degrees of freedom,

T: type of the degree of freedom value (real double, complex float, ... ),

V: Val representation of the manifold dimension of the base relation elements,

IR0, IR1, IR2, IR3: types of incidence relations with which degrees of freedom are associated in the finite element space, for each of the manifolds dimensions 0, 1, 2, 3,

F0, F1, F2, F3: types of fields with which degrees of freedom are associated in the finite element space, for each of the manifolds dimensions 0, 1, 2, 3.

Base.iterate — Function

Base.iterate(it::FEIterator, state = 1)

Advance the iterator to the next entity.

The nodes of the finite element are cached, as is a vector of all the degrees of freedom represented on the element.

Base.length — Method

Base.length(it::FEIterator)

Number of elements represented by this iterator.

Elfel.FEIterators.eldofcomps — Method

eldofcomps(it::FEIterator)

Retrieve the vector of the component numbers for each element degree of freedom.

If multiple copies of the finite element are referenced in the finite element space, each copy is referred to as component.

Elfel.FEIterators.eldofentmdims — Method

eldofentmdims(it::FEIterator)

Retrieve the vector of the entity dimensions for each element degree of freedom.

Each degree of freedom is associated with some entity of the finite element: vertices, edges, faces, and so on. This vector records the dimension of the manifold entity with which each degree of freedom is associated.

Elfel.FEIterators.eldofs — Method

eldofs(it::FEIterator)

Retrieve the vector of the element degrees of freedom

Elfel.FEIterators.elnodes — Method

elnodes(it::FEIterator)

Retrieve the vector of the nodes of the element.

Quadrature-point iterators

Assemblers

Elfel.Assemblers.AbstractSysmatAssembler — Type

AbstractSysmatAssembler

Abstract type of system-matrix assembler.

Elfel.Assemblers.AbstractSysvecAssembler — Type

AbstractSysvecAssembler

Abstract type of system vector assembler.

Elfel.Assemblers.SysmatAssemblerSparse — Type

SysmatAssemblerSparse{T<:Number} <: AbstractSysmatAssembler

Type for assembling a sparse global matrix from individual entries.

Elfel.Assemblers.SysvecAssembler — Type

SysvecAssembler

Assembler for the system vector.

Elfel.Assemblers.assemble! — Method

assemble!(self::SysmatAssemblerSparse{T}, r, c, v::T) where {T<:Number}

Assemble a single entry of a rectangular matrix.

Elfel.Assemblers.assemble! — Method

assemble!(self::SysvecAssembler{T}, i, val::T) where {T<:Number}

Assemble a single value into the row i.

Elfel.Assemblers.assemble! — Method

assemble!(self::SysvecAssembler{T}, lva) where {T<:Number}

Assemble local vector assembler.

Elfel.Assemblers.assemble! — Method

assemble!(self::SysmatAssemblerSparse{T}, lma::LocalMatrixAssembler{IT, T}) where {IT<:Integer, T<:Number}

Assemble the row numbers, column numbers, and values from a local assembler.

Elfel.Assemblers.assemble! — Method

assemble!(self::SysmatAssemblerSparse{T}, lma::Transpose{T,LocalMatrixAssembler{IT,T}}) where {IT<:Integer, T<:Number}

Assemble the row numbers, column numbers, and values from a local assembler.

Elfel.Assemblers.assemble! — Method

assemble!(self::SV, i, val::T) where {SV<:AbstractSysvecAssembler, T<:Number}

Assemble an elementwise vector.

The method assembles a column element vector using the vector of degree of freedom numbers for the rows.

Elfel.Assemblers.finish! — Method

finish!(self::SysmatAssemblerSparse)

Make a sparse matrix.

Elfel.Assemblers.finish! — Method

finish!(self::SysvecAssembler)

Make the global vector.

Elfel.Assemblers.finish! — Method

makevector!(self::SysvecAssembler)

Make the global vector.

Elfel.Assemblers.start! — Method

start!(self::SysmatAssemblerSparse{T}, nrow, ncol) where {T<:Number}

Start the assembly of a global matrix.

Elfel.Assemblers.start! — Method

start!(self::SysvecAssembler{T},  nrow::Int64) where {T<:Number}

Start assembly.

The method makes the buffer for the vector assembly. It must be called before the first call to the method assemble.

nrow= Total number of degrees of freedom.

Elfel.Assemblers.start! — Method

start!(self::SV,  nrow) where {SV<:AbstractSysvecAssembler, T<:Number}

Start assembly.

The method makes the buffer for the vector assembly. It must be called before the first call to the method assemble.

nrow= Total number of degrees of freedom.

Local Assemblers

Base.IndexStyle — Method

Base.IndexStyle(::Type{<:LocalMatrixAssembler})

The data storage is assumed to be consumed best by one-dimensional traversal through the columns in a linear fashion.

Base.IndexStyle — Method

Base.IndexStyle(::Type{<:LocalVectorAssembler})

Only linear access is provided.

Elfel.LocalAssemblers.LocalMatrixAssembler — Type

LocalMatrixAssembler{IT<:Integer, T<:Number} <: AbstractArray{T, 2}

Type of "local" matrix assembler.

Local is to be understood in the sense of in the context of a single finite element. So a local matrix is an elementwise matrix which is computed entry by entry. Then it can be assembled into the global matrix in one shot.

Elfel.LocalAssemblers.LocalVectorAssembler — Type

LocalVectorAssembler{IT<:Integer, T<:Number} <: AbstractArray{T, 1}

Type of "local" vector assembler.

Local is to be understood in the sense of in the context of a single finite element. So a local vector is an elementwise vector which is computed entry by entry. Then it can be assembled into the global vector in one shot.

Base.getindex — Method

Base.getindex(a::A, i::Int, j::Int)

Only access to a single entry of the matrix is provided.

Base.getindex — Method

Base.getindex(a::A, i::Int) where {A<:LocalVectorAssembler}

Access is provided to a single entry of the vector.

Base.setindex! — Method

Base.setindex!(a::A, v, i::Int, j::Int) where {A<:LocalMatrixAssembler}

Only access to a single entry of the matrix is provided.

Base.setindex! — Method

Base.setindex!(a::A, v, i::Int) where {A<:LocalVectorAssembler}

Access is provided to a single entry of the vector.

Base.size — Method

Base.size(a::A) where {A<:LocalMatrixAssembler}

The size is the tuple of number of rows and number of columns.

Base.size — Method

Base.size(a::A) where {A<:LocalVectorAssembler}

The size is the number of rows (entries).

Elfel.LocalAssemblers.init! — Method

init!(a::L, rdofs, cdofs) where {L<:LocalMatrixAssembler{IT, T}} where {IT, T}

Initialize the local assembler with the global degrees of freedom in the rows and columns.

The two arrays, rdofs, cdofs, define the global degree of freedom numbers for the element. The data matrix is zeroed out.

This function needs to be called for each new finite element.

Elfel.LocalAssemblers.init! — Method

init!(a::L, rdofs) where {L<:LocalVectorAssembler{IT, T}} where {IT, T}

Initialize the local assembler with the global degrees of freedom in the rows.

The array rdofs defines the global degree of freedom numbers for the element. The data vector is zeroed out.

This function needs to be called for each new finite element.

Index

Base.IndexStyle

Base.IndexStyle

Elfel.Assemblers.AbstractSysmatAssembler

Elfel.Assemblers.AbstractSysvecAssembler

Elfel.Assemblers.SysmatAssemblerSparse

Elfel.Assemblers.SysmatAssemblerSparse

Elfel.Assemblers.SysvecAssembler

Elfel.Assemblers.SysvecAssembler

Elfel.FEFields.FEField

Elfel.FEIterators.FEIterator

Elfel.FESpaces.FESpace

Elfel.FElements.FE

Elfel.FElements.FEData

Elfel.LocalAssemblers.LocalMatrixAssembler

Elfel.LocalAssemblers.LocalMatrixAssembler

Elfel.LocalAssemblers.LocalVectorAssembler

Elfel.LocalAssemblers.LocalVectorAssembler

Elfel.QPIterators.QPIterator

Elfel.QPIterators.QPIterator

Elfel.RefShapes.AbstractRefShape

Elfel.RefShapes.IntegRule

Elfel.RefShapes.RefShapeCube

Elfel.RefShapes.RefShapeInterval

Elfel.RefShapes.RefShapePoint

Elfel.RefShapes.RefShapeSquare

Elfel.RefShapes.RefShapeTetrahedron

Elfel.RefShapes.RefShapeTriangle

Base.getindex

Base.getindex

Base.getindex

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.setindex!

Base.setindex!

Base.setindex!

Base.size

Base.size

Base.size

Elfel.Assemblers.assemble!

Elfel.Assemblers.assemble!

Elfel.Assemblers.assemble!

Elfel.Assemblers.assemble!

Elfel.Assemblers.assemble!

Elfel.Assemblers.assemble!

Elfel.Assemblers.assemble!

Elfel.Assemblers.finish!

Elfel.Assemblers.finish!

Elfel.Assemblers.finish!

Elfel.Assemblers.finish!

Elfel.Assemblers.start!

Elfel.Assemblers.start!

Elfel.Assemblers.start!

Elfel.Assemblers.start!

Elfel.FEFields.datadofnums

Elfel.FEFields.datadofnums

Elfel.FEFields.dofnumtype

Elfel.FEFields.dofnumtype

Elfel.FEFields.doftype

Elfel.FEFields.doftype

Elfel.FEFields.freedofnums

Elfel.FEFields.freedofnums

Elfel.FEFields.gathersysvec!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.gathersysvec!

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestdatadofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.highestfreedofnum

Elfel.FEFields.ndofs

Elfel.FEFields.ndofs

Elfel.FEFields.ndofs

Elfel.FEFields.ndofs

Elfel.FEFields.ndofsperterm

Elfel.FEFields.ndofsperterm

Elfel.FEFields.nterms

Elfel.FEFields.nterms

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberdatadofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.numberfreedofs!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.scattersysvec!

Elfel.FEFields.setebc!

Elfel.FEFields.setebc!

Elfel.FEFields.setebc!

Elfel.FEFields.setebc!

Elfel.FEIterators.eldofcomps

Elfel.FEIterators.eldofcomps

Elfel.FEIterators.eldofentmdims

Elfel.FEIterators.eldofentmdims

Elfel.FEIterators.eldofs

Elfel.FEIterators.eldofs

Elfel.FEIterators.elnodes

Elfel.FEIterators.elnodes

Elfel.FESpaces.doftype

Elfel.FESpaces.doftype

Elfel.FESpaces.edofbfnum

Elfel.FESpaces.edofbfnum

Elfel.FESpaces.edofcompnt

Elfel.FESpaces.edofcompnt

Elfel.FESpaces.edofmdim

Elfel.FESpaces.edofmdim

Elfel.FESpaces.makeattribute

Elfel.FESpaces.makeattribute

Elfel.FESpaces.numberdofs!

Elfel.FESpaces.numberdofs!

Elfel.FESpaces.nunknowns

Elfel.FESpaces.nunknowns

Elfel.FElements.FEH1_L2

Elfel.FElements.FEH1_L2

Elfel.FElements.FEH1_Q4

Elfel.FElements.FEH1_Q4

Elfel.FElements.FEH1_T3

Elfel.FElements.FEH1_T3

Elfel.FElements.FEH1_T3_BUBBLE

Elfel.FElements.FEH1_T3_BUBBLE

Elfel.FElements.FEH1_T4

Elfel.FElements.FEH1_T6

Elfel.FElements.FEH1_T6

Elfel.FElements.FEL2_Q4

Elfel.FElements.Jacobian

Elfel.FElements.Jacobian

Elfel.FElements.Jacobian

Elfel.FElements.Jacobian

Elfel.FElements.Jacobian

Elfel.FElements.bfun

Elfel.FElements.bfun

Elfel.FElements.bfun

Elfel.FElements.bfungradpar

Elfel.FElements.bfungradpar

Elfel.FElements.bfungradpar

Elfel.FElements.jacjac

Elfel.FElements.jacjac

Elfel.FElements.jacjac

Elfel.FElements.ndofperfeat

Elfel.FElements.ndofperfeat

Elfel.FElements.ndofsperel

Elfel.FElements.ndofsperel

Elfel.FElements.ndofsperel

Elfel.FElements.ndofsperel

Elfel.FElements.ndofsperel

Elfel.FElements.nfeatofdim

Elfel.FElements.nfeatofdim

Elfel.FElements.refshape

Elfel.FElements.refshape

Elfel.FElements.shapedesc

Elfel.FElements.shapedesc

Elfel.LocalAssemblers.init!

Elfel.LocalAssemblers.init!

Elfel.LocalAssemblers.init!

Elfel.QPIterators.bfungrad

Elfel.QPIterators.weight

Elfel.RefShapes.manifdim

Elfel.RefShapes.manifdim

Elfel.RefShapes.manifdimv

Elfel.RefShapes.manifdimv

Elfel.RefShapes.quadrature

Elfel.RefShapes.quadrature

Elfel.RefShapes.quadrature

Elfel.RefShapes.quadrature

Elfel.RefShapes.quadrature

MeshCore.manifdim

MeshCore.manifdim